pyproject.toml
[tool.poetry]
name = "Travel-WebApp"
version = "0.1.0"
description = ""
authors = ["Grigoriu Ioana", "Andreea Galetan"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
aiofiles = "*"
python-jose = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
